Resolve huang and fix some memory issues (#8076)
* add compile broadcast join


* add missing dup for batch


* tmp saving for testing


* fix done channel issue


* adjust stats

* tmp saving


* fix uuid to channel mapping issue


* fix a bug will cause errCh deadlock

* fix mpool issue and channel register issue


* fix channel regist issue


* fix missing remove bug


* fix checksum issue


* remove remote addr check and some debug log


* Update join_order.go

* Update join_order.go

* reduce memory

* imporve memory

* Fix some memory use bug

* Fix confict

* Fix sca

* Fix a slip of the pen

---------

+++ pkg/sql/colexec/single/join.go
@@ -206,7 +206,7 @@ func (ctr *container) probe(bat *batch.Batch, ap *Argument, proc *process.Proces
 			sel := sels[idx]
 			for j, rp := range ap.Result {
 				if rp.Rel != 0 {
-					if err := vector.UnionOne(rbat.Vecs[j], ctr.bat.Vecs[rp.Pos], sel, proc.Mp()); err != nil {
+					if err := vector.UnionOne(rbat.Vecs[j], ctr.bat.Vecs[rp.Pos], int64(sel), proc.Mp()); err != nil {
 						rbat.Clean(proc.Mp())
 						return err
 					}
